Move ext types into own modules, add serialization test for FixedShapeTensor
