+++ serde_arrow/src/internal/schema/extensions/mod.rs
@@ -1,0 +1,6 @@
+mod fixed_shape_tensor_field;
+mod utils;
+mod variable_shape_tensor_field;
+
+pub use fixed_shape_tensor_field::FixedShapeTensorField;
+pub use variable_shape_tensor_field::VariableShapeTensorField;
